fix unhandled warning in SocketIO & missing heartbeat check, move common methods to abstract class, fix docblocks and connection inheritance
